Conflicts:
	railo-java/railo-core/src/railo/runtime/Info.ini
	railo-java/railo-core/src/railo/runtime/PageContextImpl.java
	railo-java/railo-core/src/railo/runtime/net/http/ReqRspUtil.java
	railo-java/railo-core/src/railo/transformer/bytecode/Page.java
	railo-java/railo-core/src/resource/context/railo-context.ra

+++ railo-java/railo-core/src/railo/commons/io/res/type/file/FileResource.java
@@ -169,6 +169,9 @@ public final class FileResource extends File implements Resource {
 		if(dest instanceof File) {
 			provider.lock(this);
 			try {
+				if(dest.exists() && !dest.delete())
+					throw new IOException("can't move file "+this.getAbsolutePath()+" cannot remove existing file "+dest.getAbsolutePath());
+				
 				if(!super.renameTo((File)dest)) {
 					throw new IOException("can't move file "+this.getAbsolutePath()+" to destination resource "+dest.getAbsolutePath());
 				}
